real_time_version/requirements.txt
fastapi
uvicorn[standard]
jinja2
pyzmq